+++ market14B/src/DomainLayer/backend/UserPackage/UserController.java
@@ -85,7 +85,7 @@ public class UserController {
         return addToRegUserMap(reg);
     }
 
-    public int Buy(String username) throws Exception {
+    public double Buy(String username) throws Exception {
         User user = getUser(username);
         if (user != null)
             return user.Buy();
